Refactored WebViewContentHelper.
Uses common base class implementation to provide uniform exceptions, reflection use.
