+++ src/MobileKidsIdApp/MobileKidsIdApp.Droid/MobileKidsIdApp.Droid.csproj
@@ -199,6 +199,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\ContactPicker.cs" />
     <Compile Include="Services\PhotoPicker.cs" />
+    <Compile Include="Services\WebViewContentHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
