+++ src/MobileKidsIdApp/MobileKidsIdApp.Droid/Services/WebViewContentHelper.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.IO;
+using System.Reflection;
+using MobileKidsIdApp.Services;
+
+[assembly: Xamarin.Forms.Dependency(typeof(MobileKidsIdApp.Droid.Services.WebViewContentHelper))]
+namespace MobileKidsIdApp.Droid.Services
+{
+    public class WebViewContentHelper : WebViewContentHelperBase
+    {
+        public override string GetBaseUrl()
+        {
+            return "file:///android_asset/";
+        }
+    }
+}
