+++ src/MobileKidsIdApp/MobileKidsIdApp.iOS/MobileKidsIdApp.iOS.csproj
@@ -101,7 +101,7 @@
     <None Include="packages.config">
       <SubType>Designer</SubType>
     </None>
-    <Compile Include="Services\WebViewContentHelper_IOS.cs" />
+    <Compile Include="Services\WebViewContentHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MobileKidsIdApp.Models\MobileKidsIdApp.Models.csproj">
